pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lerobot-ros2"
version = "0.1.0"
description = "ROS2 integration bridge for LeRobot: End-to-end robot learning pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["robotics", "robot learning", "ros2", "lerobot", "imitation learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Robotics",
]

dependencies = [
    "numpy>=1.21.0,<2.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    # NOTE: lerobot is NOT a dependency! Only used as reference in references/lerobot/
    "opencv-python>=4.5.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.60.0",
    "pyarrow>=10.0.0",
    "pandas>=1.3.0",
    "rosbags>=0.9.0",
    "transforms3d>=0.4.1",
    "huggingface-hub>=0.19.0",
    "dynamixel-sdk>=3.7.0",
    "feetech-servo-sdk>=1.0.0",  # Feetech motors (SO-100, SO-101)
    "pyserial>=3.5",
    "pygame>=2.0.0",  # For calibration/control GUIs
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

ros2 = [
    # ROS2 packages are typically installed via apt, not pip
    # This is a placeholder for documentation purposes
    # Install: sudo apt install ros-humble-ros-base ros-humble-cv-bridge
]

realsense = [
    "pyrealsense2>=2.50.0",
]

isaac-sim = [
    # Isaac Sim dependencies
    # Installed via NVIDIA Isaac Sim package
]

all = [
    "lerobot-ros2[dev,realsense]",
]

[project.urls]
Homepage = "https://github.com/yourusername/lerobot-ros2"
Documentation = "https://github.com/yourusername/lerobot-ros2/docs"
Repository = "https://github.com/yourusername/lerobot-ros2"
"Bug Tracker" = "https://github.com/yourusername/lerobot-ros2/issues"

[project.scripts]
lerobot-ros2-collect = "lerobot_ros2.data_collection.cli:collect_main"
lerobot-ros2-train = "lerobot_ros2.training.cli:train_main"
lerobot-ros2-deploy = "lerobot_ros2.deployment.cli:deploy_main"
lerobot-ros2-convert = "lerobot_ros2.training.cli:convert_main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["lerobot_ros2*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose --cov=lerobot_ros2 --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
